added method that reads the vending mechine cvs to populate sales report map, updated candy,drinks,chips,and gum to use big decimal in constructor but will have error until inventory constructor is changed. started working on update method for sales report that should read the log and write the report to a file.

+++ src/main/java/com/techelevator/products/Candy.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 public class Candy extends Inventory {
 
-    public Candy(String vendPosition, String productName, double price, String type) {
+    public Candy(String vendPosition, String productName, BigDecimal price, String type) {
         super(vendPosition, productName, price, type);
     }
 }
